Configured Spring Security Authentication and role based authorisation for Admins, General Users and Users

+++ src/main/java/com/CinemaHub/CinemaHub/Repository/ShowRepo.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ShowRepo extends JpaRepository<MovieShow,Long> {
+public interface ShowRepo extends JpaRepository<MovieShow,Integer> {
 
 
     @Query(value = "SELECT " +
